+++ clients/src/main/java/io/strimzi/common/records/http/producer/ProducerRecord.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.http.producer;
+package io.strimzi.common.records.http.producer;
 
 import io.strimzi.test.tracing.HttpContext;
